refactored to reduce code duplication for tube rack scan files
